// layer_1/erp2/frontend/src/store/enums/storeEnums.ts
enum Mutations {
  CLEAR_USER_LIST = "clearUserList",
  SET_AUTH_STATUS = "setAuthenticatedStatus",
  SET_BC_AUTH_STATUS = "setBcAuthStatus",
  SET_COMPANY_BRAND = "setCompanyName",
  SET_COMPANY_DETAILS = "setCompanyDetails",
  SET_CUSTOMERS = "setCustomerList",
  SET_INBOUND = "setInbound",
  SET_MAX_BLOCK_NUMBER = "setMaxBlockNumber",
  SET_MFA = "setMFA",
  SET_ORDER = "setOrder",
  SET_OUTBOUND = "setOutbound",
  SET_PASSWORD = "setPassword",
  SET_PRODUCT_TO_BUY = "setProductToBuy",
  SET_PRODUCT_TO_SELL = "setProductToSell",
  SET_RESOURCE = "setResource",
  SET_REFRESH_TOKEN = "setRefreshTokenToken",
  SET_STATS_BC_VISIBILITY = "setBcStatsVisibility",
  SET_STATS_DB_VISIBILITY = "setDbStatsVisibility",
  SET_STATS_VISIBILITY = "setStatisticsVisibility",
  SET_STATUS_CODE = "setStatusCode",
  SET_TOTAL_USERS = "setUsersCounter",
  SET_USER = "setUser",
  SET_USER_LIST = "setUserList",
  //TOKEN
  SET_REPEAT_MODE = "setRepeatMode",
  SET_PEDIGREE_TOKEN = "setPedigreeToken",
  SET_PROCESS_INFO = "setProcessInfo",
  SET_PROCESS_LIST = "addProcessToList",
  SET_PROCESS_TOKEN = "setProcessToken",
  SET_PRODUCT_INFO = "setProductInfo",
  SET_PRODUCT_TOKEN = "setProductToken",
  SET_STAKEHOLDER_TOKEN = "setStakeholder",
  SET_STAKEHOLDER_INFO = "setStakeholderInfo",
  SET_TRACE_INFO = "setTraceInfo",
  SET_TRACE_TOKEN = "setTraceToken",
  UNSET_PROCESS_INFO = "clearProcessInfo",
  UNSET_PROCESS_LIST = "clearProcessList",
  UNSET_PROCESS_TOKEN = "clearProcessToken",
  UNSET_PRODUCT_INFO = "clearProductInfo",
  UNSET_PRODUCT_TOKEN = "clearProductToken",
  UNSET_STAKEHOLDER = "clearStakeholder",
  UNSET_STAKEHOLDER_INFO = "clearStakeholderInfo",
  UNSET_TRACE_TOKEN = "clearTraceToken",
  UNSET_TRACE_TOKEN_INFO = "clearTraceTokenInfo",
}